requirements.txt
Django==4.2
django-jazzmin==2.6.0
inertia-django==0.5.2
django-vite==2.0.2
django-tagify==2-1.0.19